Removed redux and replace it with React hooks instead.
