+++ diamonds-viewer/package.json
@@ -7,11 +7,7 @@
     "lodash": "^4.17.11",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
-    "react-redux": "^7.0.1",
     "react-scripts": "2.1.8",
-    "redux": "^4.0.1",
-    "redux-promise": "^0.6.0",
-    "redux-thunk": "^2.3.0",
     "styled-components": "^4.2.0"
   },
   "scripts": {
