--- diamonds-viewer/src/actions/index.js
@@ -1,2 +1,0 @@
-export const UPDATE_BOARD = "update_board";
-export const IS_FETCHING_BOARD = "is_fetching_board";
