+++ diamonds-viewer/src/hooks/index.js
@@ -1,4 +1,5 @@
 import useInterval from "./useInterval";
 import useFetchRepeatedly from "./useFetchRepeatedly";
+import useBoard from "./useBoard";
 
-export { useInterval, useFetchRepeatedly };
+export { useInterval, useFetchRepeatedly, useBoard };
