+++ diamonds-viewer/src/hooks/useFetchRepeatedly.js
@@ -2,12 +2,17 @@ import { useState, useEffect } from "react";
 import axios from "axios";
 import useInterval from "./useInterval";
 
-export default (url, delay, initialValue = null) => {
-  let [response, setResponse] = useState(initialValue);
+export default (url, delay, baseResponse = null) => {
+  let [response, setResponse] = useState(baseResponse);
+  let [isFetching, setFetching] = useState(false);
 
   const fetch = async () => {
-    const { data } = await axios.get(url);
-    setResponse(data);
+    if (!isFetching) {
+      setFetching(true);
+      const { data } = await axios.get(url);
+      setResponse(data);
+      setFetching(false);
+    }
   };
 
   useEffect(() => {
