--- diamonds-viewer/src/reducers/index.js
@@ -1,8 +1,0 @@
-import { combineReducers } from "redux";
-import boardReducer from "./boardReducer";
-import botsReducer from "./botsReducer";
-
-export default combineReducers({
-  board: boardReducer,
-  bots: botsReducer
-});
